+++ azure_monitor/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- Remove dependency metrics from auto-collection
+- ([#92](https://github.com/microsoft/opentelemetry-azure-monitor-python/pull/92))
+
 ## 0.3b.1
 Released 2020-05-21